+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Ensure that `SymbolEntryIndexDiff` correctly handles differences in both members as well as types.
 * Add `SymbolEntry.isType()` and `SymbolEntry.isMember()` helper methods.
 * Add `SymbolEntryIndexDiff.printToString()` to display differences in human readable form.
+* Add `SymbolEntryIndexDiff.hasDifferences()` to determine if there is any differences.
 
 ### [v0.04](https://github.com/realityforge/gwt-symbolmap/tree/v0.04) (2018-06-05)
 [Full Changelog](https://github.com/realityforge/gwt-symbolmap/compare/v0.03...v0.04)